Move structs to correct files
